refactor: extract public and internal wallet clients
This removes the need for a public and an internal issuance agent.
